--- src/main/java/com/mindDiary/mindDiary/exception/EmailDuplicatedException.java
@@ -1,9 +1,0 @@
-package com.mindDiary.mindDiary.exception;
-
-public class EmailDuplicatedException extends BusinessException {
-
-  @Override
-  public ErrorCode getErrorCode() {
-    return ErrorCode.EMAIL_DUPLICATED;
-  }
-}
