--- src/main/java/com/mindDiary/mindDiary/exception/InvalidEmailTokenException.java
@@ -1,8 +1,0 @@
-package com.mindDiary.mindDiary.exception;
-
-public class InvalidEmailTokenException extends BusinessException {
-  @Override
-  public ErrorCode getErrorCode() {
-    return ErrorCode.INVALID_EMAIL_TOKEN;
-  }
-}
