--- src/main/java/com/mindDiary/mindDiary/exception/NicknameDuplicatedException.java
@@ -1,8 +1,0 @@
-package com.mindDiary.mindDiary.exception;
-
-public class NicknameDuplicatedException extends BusinessException {
-  @Override
-  public ErrorCode getErrorCode() {
-    return ErrorCode.NICKNAME_DUPLICATED;
-  }
-}
