--- src/main/java/com/mindDiary/mindDiary/exception/NotMatchedIdException.java
@@ -1,8 +1,0 @@
-package com.mindDiary.mindDiary.exception;
-
-public class NotMatchedIdException extends BusinessException {
-  @Override
-  public ErrorCode getErrorCode() {
-    return ErrorCode.NOT_MATCHED_ID;
-  }
-}
