--- src/main/java/com/mindDiary/mindDiary/exception/NotMatchedPasswordException.java
@@ -1,8 +1,0 @@
-package com.mindDiary.mindDiary.exception;
-
-public class NotMatchedPasswordException extends BusinessException {
-  @Override
-  public ErrorCode getErrorCode() {
-    return ErrorCode.NOT_MATCHED_PASSWORD;
-  }
-}
