+++ src/main/java/com/mindDiary/mindDiary/service/UserServiceImpl.java
@@ -2,11 +2,11 @@ package com.mindDiary.mindDiary.service;
 
 import com.mindDiary.mindDiary.dto.TokenDTO;
 import com.mindDiary.mindDiary.dto.UserDTO;
-import com.mindDiary.mindDiary.exception.EmailDuplicatedException;
-import com.mindDiary.mindDiary.exception.NicknameDuplicatedException;
-import com.mindDiary.mindDiary.exception.InvalidEmailTokenException;
-import com.mindDiary.mindDiary.exception.NotMatchedIdException;
-import com.mindDiary.mindDiary.exception.NotMatchedPasswordException;
+import com.mindDiary.mindDiary.exception.businessException.EmailDuplicatedException;
+import com.mindDiary.mindDiary.exception.businessException.NicknameDuplicatedException;
+import com.mindDiary.mindDiary.exception.businessException.InvalidEmailTokenException;
+import com.mindDiary.mindDiary.exception.businessException.NotMatchedIdException;
+import com.mindDiary.mindDiary.exception.businessException.NotMatchedPasswordException;
 import com.mindDiary.mindDiary.repository.UserRepository;
 import com.mindDiary.mindDiary.strategy.email.EmailStrategy;
 import com.mindDiary.mindDiary.strategy.jwt.TokenStrategy;
